requirements.txt
PyYAML
git+https://github.com/nlevitt/surt.git@py3
git+https://github.com/nlevitt/youtube-dl.git@brozzler
git+https://github.com/seomoz/reppy.git   # https://github.com/seomoz/reppy/commit/7661606c not in pypi package
requests
git+https://github.com/nlevitt/websocket-client.git@tweaks
rethinkdb
